backend/requirements.txt
numpy==1.26.4
openai==1.35.3
streamlit==1.36.0
streamlit-chat==0.1.1
streamlit-feedback==0.1.3
pinecone==4.0.0 
psycopg2
python-dotenv